Database Revision
Revised the database to increase performance and to hold data in a manner consistant with its use. Also changed the website to remove hardcoding of several variables related to the database.

This should allow the site to be  automated in terms of handling future NaNoWriMo seasons, and allow the front end to match whatever the back end is, allowing a single point of change to update the website.